requirements.txt
pydicom>=2.4.0
SimpleITK>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
torch>=2.0.0
monai>=1.3.0
pyradiomics>=3.0.1
